# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Student Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface UploadUrlRequest {
  /**
     * Original file name.
     * @minLength 1
     */
  name: string;
  /**
     * File size in bytes.
     * @minimum 1
     */
  size: number;
  /**
     * MIME type of the file (e.g. `image/jpeg`).
     * @minLength 1
     */
  contentType: string;
}

export interface UploadUrlResponse {
  /**
     * Presigned GCS URL for PUT upload.
     * @minLength 1
     */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type Gender = typeof Gender[keyof typeof Gender];


export const Gender = {
  Male: 'Male',
  Female: 'Female',
  Other: 'Other',
} as const;

export interface Student {
  id: number;
  admissionNumber: string;
  name: string;
  course: string;
  year: number;
  dateOfBirth: string;
  email: string;
  mobileNumber: string;
  gender: Gender;
  address: string;
  /** @nullable */
  photoUrl: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface StudentInput {
  /**
     * @minLength 1
     * @maxLength 150
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  course: string;
  /**
     * @minimum 1
     * @maximum 8
     */
  year: number;
  dateOfBirth: string;
  /**
     * @minLength 3
     * @maxLength 150
     */
  email: string;
  /**
     * @minLength 7
     * @maxLength 20
     */
  mobileNumber: string;
  gender: Gender;
  /**
     * @minLength 1
     * @maxLength 500
     */
  address: string;
  /** @nullable */
  photoUrl?: string | null;
}

export interface StudentUpdate {
  /**
     * @minLength 1
     * @maxLength 150
     */
  name?: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  course?: string;
  /**
     * @minimum 1
     * @maximum 8
     */
  year?: number;
  dateOfBirth?: string;
  /**
     * @minLength 3
     * @maxLength 150
     */
  email?: string;
  /**
     * @minLength 7
     * @maxLength 20
     */
  mobileNumber?: string;
  gender?: Gender;
  /**
     * @minLength 1
     * @maxLength 500
     */
  address?: string;
  /** @nullable */
  photoUrl?: string | null;
}

export interface StudentListResponse {
  data: Student[];
  total: number;
  page: number;
  pageSize: number;
}

export interface CountBucket {
  label: string;
  count: number;
}

export interface StudentAnalytics {
  totalStudents: number;
  byCourse: CountBucket[];
  byYear: CountBucket[];
  byGender: CountBucket[];
}

export interface ActivityLog {
  id: number;
  action: string;
  entityType: string;
  /** @nullable */
  entityId: number | null;
  description: string;
  createdAt: string;
}

export interface ActivityLogListResponse {
  data: ActivityLog[];
  total: number;
  page: number;
  pageSize: number;
}

export type ListStudentsParams = {
/**
 * Free-text search across name, admission number, and email
 */
search?: string;
course?: string;
year?: number;
gender?: Gender;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: number;
sortBy?: ListStudentsSortBy;
sortOrder?: ListStudentsSortOrder;
};

export type ListStudentsSortBy = typeof ListStudentsSortBy[keyof typeof ListStudentsSortBy];


export const ListStudentsSortBy = {
  name: 'name',
  admissionNumber: 'admissionNumber',
  course: 'course',
  year: 'year',
  createdAt: 'createdAt',
} as const;

export type ListStudentsSortOrder = typeof ListStudentsSortOrder[keyof typeof ListStudentsSortOrder];


export const ListStudentsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type ListActivityLogsParams = {
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: number;
};